+++ extensions/processors/ChannelReductor/ChannelReductor.hpp
@@ -36,8 +36,8 @@ class ChannelReductor : public IProcessor {
 
   protected:
     // Data ports
-    PortIn<MultiChannelType<float>>  *data_in_port_;
-    PortOut<MultiChannelType<float>> *data_out_port_;
+    PortIn<MultiChannelType<double>>  *data_in_port_;
+    PortOut<MultiChannelType<double>> *data_out_port_;
 
     // Options
     options::Int n_messages_{-1};
